Abilities

+++ examples/basic_todo/app/controller_actions/todos_controller/destroy.rb
@@ -6,6 +6,7 @@ class TodosController::Destroy < BaseAction
   def perform
     return Home::ShowRedirect.build unless current_user
     return Todos::NotFoundView.build(todo_id: params[:id]) unless todo
+    return Defaults::ForbiddenView.build unless current_ability.can?(:destroy, todo)
 
     todo.destroy
